cleaner tests structure
