+++ tests/mocked_browser/conftest.py
@@ -1,6 +1,5 @@
 """Configuration for pytest runner."""
-
-pytest_plugins = 'pytester'
+import mock
 
 import pytest
 
@@ -11,7 +10,27 @@ def splinter_session_scoped_browser():
     return False
 
 
-@pytest.fixture(autouse=True)
-def mocked_browser(mocked_browser):
-    """Make mocked browser fixture autoused."""
-    return mocked_browser
+@pytest.yield_fixture(autouse=True)
+def mocked_browser(browser_pool, request):
+    """Mock splinter browser."""
+    # to avoid re-using of cached browser from other tests
+    for browser in browser_pool.values():
+        browser.quit()
+    browser_pool.clear()
+
+    def mocked_browser(*args, **kwargs):
+        mocked_browser = mock.MagicMock()
+        mocked_browser.driver = mock.MagicMock()
+        mocked_browser.driver.profile = mock.MagicMock()
+
+        def save_screenshot(path):
+            with open(path, 'w'):
+                pass
+
+        mocked_browser.driver.save_screenshot = save_screenshot
+        return mocked_browser
+
+    patcher = mock.patch('pytest_splinter.plugin.splinter.Browser', mocked_browser)
+    patcher.start()
+    yield mocked_browser
+    patcher.stop()
