+++ tests/mocked_browser/test_screenshot.py
